% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_weight_boot.R
\name{dist_weight_boot}
\alias{dist_weight_boot}
\title{Perform a bootstrap likelihood ratio test on a scalescape object}
\usage{
dist_weight_boot(
  mod.full,
  mod.reduced,
  data = NULL,
  nboot = 2000,
  plot.fits = TRUE,
  verbose = FALSE,
  pb.flag = TRUE,
  n.breaks = NULL,
  optim.method = "L-BFGS-B",
  lower = NULL,
  upper = NULL,
  n.partition = 10
)
}
\arguments{
\item{mod.full}{a \code{scalescape} object; the landscape model fitted by \code{dist_weight}.}

\item{mod.reduced}{the corresponding "local" model object, without landscape variables.}

\item{data}{a data frame with local and landscape predictors and response variables.
By default, this is extracted from \code{mod.full}.}

\item{nboot}{the number of bootstraps. To determine whether a variable is significant
at the alpha level 0.05, \code{nboot=2000} should be used (default).}

\item{plot.fits}{produce histograms of the bootstrapped range and deviance values (default = TRUE).}

\item{verbose}{produce output for each iteration of the bootstrap (default = FALSE)}

\item{pb.flag}{show a progress bar for completion of the bootstraps (default = TRUE)}

\item{n.breaks}{specify...}

\item{optim.method}{specify method for optimization (default = \code{"L-BFGS-B"}).
See \code{optim()}.}

\item{lower}{lower bound on the variables for the \code{"L-BFGS-B"} method. See \code{optim()}.}

\item{upper}{upper bound on the variables for the \code{"L-BFGS-B"} method. See \code{optim()}.}

\item{n.partition}{number of partitions to divide the log-likelihood profile, in order
to avoid identifying false maxima.}
}
\value{
\code{dist_weight_boot} returns an object of class \code{scalescape.boot}. This is a list
containing the following:
\itemize{
  \item the full model
  \item the reduced model
  \item the deviance value of the full vs. reduced model
  \item the mean deviance value of the bootstraps
  \item the standard deviation of the bootstrap deviance values
  \item the P value for the bootstrap likelihood ratio test
  \item a data frame of log-likelihood and deviance values for each iteration of the bootstrap
  \item a data frame of coefficient values for each iteration of the bootstrap
  }
}
\description{
\code{dist_weight_boot} conducts a parametric bootstrap test of the null hypothesis that one or
more of the landscape predictors in a model fit using \code{dist_weight} has no
effect on the response variable.
}
\details{
\code{dist_weight_boot} simulates datasets using the attributes and parameter values
from the model specified with dist_weight() and then re-fits the full and reduced models
to all the datasets using optim() to maximize the range parameter with each iteration.
It then compares them to test the null hypothesis that the reduced model fits better
than the full model, producing a p-value for the effect of the additional parameters in
the full model on the response. Although \code{dist_weight} produces p-values for the
regression coefficients, these p-values are conditional on the estimate of the range
parameter, and consequently they will likely have inflated type I error rates. By
bootstrapping, \code{dist_weight_boot} accounts for the co-dependence of regression
coefficient and range parameter. Therefore, p-values reported for landscape predictor(s)
should come from \code{dist_weight_boot} rather than \code{dist_weight}.
}
