% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_matrix.R
\name{landscape_matrix}
\alias{landscape_matrix}
\title{Create a matrix for the landscape surrounding each site}
\usage{
landscape_matrix(raster, sites, max.radius, is.factor = FALSE)
}
\arguments{
\item{raster}{A raster object associated with the landscape variable of interest.}

\item{sites}{An inventory of sites formatted as a 2-column, X-Y data frame,
sf object, or sp object.}

\item{max.radius}{The maximum radius you wish to evaluate. Must be less than the
maximum extent of the raster.}
}
\value{
A matrix used by \code{dist_weight} to do the weighting and fit the model.
}
\description{
\code{landscape_matrix} creates the matrix that is used by \code{dist_weight}
 to weight the landscape variable and fit the model.
}
\details{
Each row of the \code{landscape_matrix} output corresponds to a raster pixel at a given
location around a site, and each column corresponds to a site. The column
'dist' gives the distance of each pixel to the focal site. For binary
classifications, cell values in the matrix represent whether that pixel is
filled by a land cover type of interest (1 if yes, 0 if no), and for continuous
classifications they correspond to the value of the landscape variable of
interest. When applying \code{landscape_matrix} to a raster, you will need to specify:
\itemize{
   \item The raster associated with the landscape variable you are interested in.
   \item A list of sites formatted as a 2-column X-Y data frame, \code{sf} object,
   or \code{sp} object.
   \item The maximum radius you wish to evaluate.This radius should be larger than
   what you believe to be the maximum relevant spatial scale for your response variable
   (although the larger it is, the longer the computation time). All raster cells up to
   this distance will be included in the parameter optimization process used to define
   the range parameter. The value for \code{max.radius} must fall within the bounds of
   the raster. If \code{max.radius} around at least one site falls outside the bounds of
   your raster, you will receive an error message indicating this problem and should
   select a raster with a greater extent or choose a smaller \code{max.radius}.
}
}
