% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.scalescape.R
\name{plot.scalescape}
\alias{plot.scalescape}
\title{Plotting scalescape objects}
\usage{
\method{plot}{scalescape}(x, ...)
}
\arguments{
\item{x}{an object of class \code{scalescape} or \code{scalescape.boot}}

\item{...}{arguments to be passed to methods}
}
\description{
\code{plot} method for classes \code{scalescape} and \code{scalescape.boot}. For objects of
class \code{scalescape}, \code{summary.scalescape} returns two plots:
\enumerate{
  \item the log-likelihood values for the range parameter up to the distance (in meters),
   specified as the \code{max.radius} in \code{landscape_matrix}.
   \item the shape of the weighting function, i.e.how the weight of  the landscape predictor
   decreases with distance (in meters), given the maximum likelihood range value.
 }
For objects of class \code{scalescape.boot}, \code{summary.scalescape} returns two histograms:
\enumerate{
  \item the frequency with which different range parameter values were identified as optimal
  \item the distribution of deviance values generated by the bootstraps, from which the p-value is derived
}
}
