% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalescape.data.R
\docType{data}
\name{scalescape.data}
\alias{scalescape.data}
\title{Bee abundance and floral data}
\format{
A \code{data.frame} 18 observations of 6 variables
\describe{
  \item{**site.id**}{unique site identifier}
  \item{**lon**}{longitude}
  \item{**lat**}{latitude}
  \item{**sqrt.abundance**}{square root of wild bee abundance counts}
  \item{**floral.index**}{index of local floral abundance at sampling sites}
  \item{**temp**}{mean temperature at each site at time of sampling}
  }
}
\source{
Erin Lowe
}
\usage{
data(scalescape.data)
}
\description{
Data from 2018 field sampling of wild bees
}
\keyword{datasets}
