% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalescape.sites.R
\docType{data}
\name{scalescape.sites}
\alias{scalescape.sites}
\title{Sampling locations}
\format{
An object of class \code{SpatialPointsDataFrame} with coordinates of sampling locations
}
\source{
Erin Lowe
}
\usage{
data(scalescape.sites)
}
\description{
Sampling locations of wild bees. Example from Lowe et al. 2021.
}
\keyword{datasets}
