% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.scalescape.R
\name{simulate.scalescape}
\alias{simulate.scalescape}
\title{Simulate a distance-weighted model of landscape effects on an environmental response}
\usage{
\method{simulate}{scalescape}(
  mod.full,
  mod.reduced,
  new.coef,
  new.range,
  new.sigma = NULL,
  data = NULL,
  nsim = 1
)
}
\arguments{
\item{mod.full}{the full (landscape) model upon which to base the simulations}

\item{mod.reduced}{the reduced (local) model}

\item{new.coef}{simulated coefficient for the landscape effect}

\item{new.range}{simulated range parameter (i.e. scale of effect) for the landscape effect}

\item{new.sigma}{simulated standard deviation. By default, this is extracted from \code{mod.reduced}}

\item{data}{a data frame with local and landscape predictors and response variables.
By default, this is extracted from \code{mod.full}}

\item{nsim}{number of simulations (default = 1)}
}
\value{
\code{simulate.scalescape} returns a data frame of predictor and simulated response variables.
}
\description{
\code{simulate} method for class \code{scalescape}.Currently only available for linear models
}
