% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.scalescape.R
\name{summary.scalescape}
\alias{summary.scalescape}
\title{Summarizing scalescape model fits}
\usage{
\method{summary}{scalescape}(object, ...)
}
\arguments{
\item{object}{an object of class \code{scalescape} or \code{scalescape.boot}}
}
\value{
For objects of class \code{scalescape},  \code{summary.scalescape} returns a list of
summary statistics of the fitted model given in \code{object}, including:
\itemize{
   \item the estimated range parameter, in meters
   \item metrics of model fit (log-likelihood, AIC, BIC), and the number of parameters
   \item the call for the local (no landscape variables) model
   \item the call for the distance-weighted landscape model
   \item residuals
   \item coefficients, standard errors, t-statistics and corresponding p-values
}
 For objects of class \code{scalescape.boot},  \code{summary.scalescape} returns a list of
 summary statistics of the bootstrap likelihood ratio test comparing the full and reduced
 models, including:
\itemize{
  \item the number of bootstrapped datasets successfully refit
  \item the call for the full (e.g., landscape) model
  \item the call for the reduced (e.g., local) model
  \item the observed deviance, bootstrap deviance, standard deviation, and corresponding p-value
  }
}
\description{
\code{summary} method for classes \code{scalescape} and \code{scalescape.boot}
}
\details{
Note that although \code{summary.scalescape} displays p-values for the regression coefficients
(given by the underlying \code{lm()},\code{glm}, \code{lmer}, \code{glmer}, \code{lme},
or \code{gls} functions, p-values for landscape effets are conditional on the estimate
of the range parameter, and consequently they will likely have inflated type I error
rates. The \code{dist_weight_boot()} function uses a bootstrap likelihood ratio test to
generate a single p-value for the landscape predictor variable(s) in the model. By
bootstrapping, it accounts for the co-dependence of regression coefficient and range
parameter. Therefore, p-values reported for landscape predictor(s) should come from
\code{dist_weight_boot} rather than \code{dist_weight}.
}
