% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_weight.R
\name{dist_weight}
\alias{dist_weight}
\title{Fit a distance-weighted model of landscape effects on an environmental response}
\usage{
dist_weight(
  mod0,
  landscape.vars,
  landscape.formula,
  data = NULL,
  weight.fn = "Gaussian",
  plot.fits = TRUE,
  lower = NULL,
  upper = NULL,
  init.range = NULL,
  n.partition = 10,
  opt.range = NULL,
  optim.method = "L-BFGS-B"
)
}
\arguments{
\item{mod0}{a "local" model object, without landscape variables.}

\item{landscape.vars}{list of names landscape matrices (one for each landscape variable).}

\item{landscape.formula}{formula containing the landscape predictors.}

\item{data}{a data frame with local predictors and response variables, where
each row as a site sorted in the same order as the original site data frame.}

\item{weight.fn}{the type of weighting function to use; "Gaussian" (default) or "exponential".}

\item{plot.fits}{specify whether to produce plots (default = TRUE).}

\item{lower}{lower bound on the variables for the \code{"L-BFGS-B"} method. See \code{optim}.}

\item{upper}{upper bound on the variables for the \code{"L-BFGS-B"} method. See \code{optim}.}

\item{init.range}{starting point distance for fitting the range parameter.}

\item{n.partition}{number of partitions to divide the log-likelihood profile, in order
to avoid identifying false maxima.}

\item{opt.range}{specified value range parameter value. If specified, \code{dist_weight}.
will fit the model using that value as the value of the range parameter rather than using.
\code{optim} to identify the value that maximizes the log likelihood.}

\item{optim.method}{specify method for optimization (default = \code{"L-BFGS-B"}).
See \code{optim}.}
}
\value{
\code{dist_weight} returns an object of class \code{scalescape}. This is a list
containing the following:
\itemize{
   \item \code{opt.range} the estimate of the range value
   \item \code{logLik} the maximum likelihood value
   \item \code{AIC} AIC value for model fit
   \item \code{BIC} BIC value for model fit
   \item \code{npar} the number of model parameters
   \item \code{data} the original (local) data frame used to fit the model
   \item \code{coef} the coefficient of the landscape variable(s)
   \item \code{mod} the landscape model, with distance-weighted landscape effect
   \item \code{mod0} the local model, without landscape effects
   \item \code{landscape.formula} the formula specified to fit the landscape model
   \item \code{data.with.landscape} a data frame that contains the original local variable and distance-weighted landscape variable(s)
   \item \code{landscape.vars} the list of landscape matrices for each landscape variable
   \item \code{weight.fn} the specified weighting function
   \item \code{max.Dist} the specified maximum distance
   }
}
\description{
\code{dist_weight} fits a specified model containing local and landscape variables.
}
\details{
\code{dist_weight} fits the model using the function \code{optim} (method L-BFGS-B)
to find the values of the model parameters that maximize the log-likelihood of the
model fit to the data. It can be run with many types of regression models in R:
\code{lm} and \code{glm} in base R, \code{lmer} and \code{glmer} in \code{lme4},
and \code{lme} and \code{gls} in \code{nlme.} It adopts the model syntax of the
specified regression model,making it easy to use models of any type.Although
\code{dist_weight} producesp-values for the regression coefficients (given by the
underlying \code{lm},\code{glm}, \code{lmer}, \code{glmer}, \code{lme}, or
\code{gls} functions,these p-values are conditional on the estimate of the range
parameter, and consequently they will likely have inflated type I error rates. The
\code{dist_weight_boot}function uses a bootstrap likelihood ratio test to generate
a single p-value for the landscape predictor variable(s) in the model. By
bootstrapping, it accounts for the co-dependence of regression coefficient and
range parameter. Therefore, p-values reported for landscape predictor(s) should come from
\code{dist_weight_boot} rather than \code{dist_weight}.
}
